Allow proxy to be set through options
This is useful if you are using the gem on a host with a dynamic ip, since access to the google api is limited by a whitelist ip-list. This makes it possible to proxy only requests for freebase instead of having to set any of the proxy ENV-variables which could potentially cause other requests to also be proxied.